requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
asgiref==3.11.0
    # via django
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
colorama==0.4.6
    # via tqdm
distro==1.9.0
    # via openai
django==5.2.8
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.16.1
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   openai
    #   python-telegram-bot
idna==3.11
    # via
    #   anyio
    #   httpx
jiter==0.12.0
    # via openai
openai==2.8.1
    # via -r requirements.in
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
python-telegram-bot==22.5
    # via -r requirements.in
sniffio==1.3.1
    # via openai
sqlparse==0.5.4
    # via django
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via django
